Related work items: #143351

+++ Source/ail/Source/Library/AppInfra/src/main/java/com/philips/platform/appinfra/rest/RestManager.java
@@ -148,7 +148,7 @@ public class RestManager implements RestInterface {
             }
         }
 
-        return cacheSize!=null?cacheSize.intValue():DiskBasedCache.DEFAULT_DISK_USAGE_BYTES;
+        return cacheSize != null ? cacheSize : DiskBasedCache.DEFAULT_DISK_USAGE_BYTES;
     }
 
     private class ServiceIDResolver implements HurlStack.UrlRewriter {
